# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  model.GetAllStudentsResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/model.Student'
        type: array
      status:
        type: string
    type: object
...
  model.GetStudentByIDResponse:
    properties:
      data:
        $ref: '#/definitions/model.Student'
      status:
        type: string
    type: object
...
  model.Student:
    properties:
      academic_year:
        type: string
      advisor_id:
        type: string
      created_at:
        type: string
      full_name:
        type: string
      id:
        type: string
      program_study:
        type: string
      student_id:
        type: string
      user_id:
        type: string
    type: object
...
host: localhost:3001
info:
  contact: {}
  description: API untuk sistem pelaporan prestasi mahasiswa dengan dukungan multi-role
    (Admin, Dosen Wali, Mahasiswa) dan field prestasi yang fleksibel
  title: Sistem Pelaporan Prestasi Mahasiswa API
  version: "1.0"
paths:
...
  /reports/student:
    get:
      consumes:
      - application/json
      description: Mengambil laporan prestasi student yang sedang login
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Get current student report
      tags:
      - Reports
  /reports/student/{id}:
    get:
      consumes:
      - application/json
      description: Mengambil laporan prestasi student berdasarkan ID
      parameters:
      - description: Student ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Get student report by ID
      tags:
      - Reports
...
  /students:
    get:
      consumes:
      - application/json
      description: Mengambil daftar semua student dari database. Hanya dapat diakses
        oleh admin dengan permission user:manage
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.GetAllStudentsResponse'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Get all students
      tags:
      - Students
  /students/{id}:
    get:
      consumes:
      - application/json
      description: Mengambil data student spesifik berdasarkan ID. Hanya dapat diakses
        oleh admin dengan permission user:manage
      parameters:
      - description: Student ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.GetStudentByIDResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Get student by ID
      tags:
      - Students
  /students/{id}/achievements:
    get:
      consumes:
      - application/json
      description: Mengambil daftar achievements milik student tertentu dengan pagination.
        Dapat diakses dengan permission achievement:read
      parameters:
      - description: Student ID
        in: path
        name: id
        required: true
        type: string
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Limit per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Get student achievements
      tags:
      - Students
  /students/{id}/advisor:
    put:
      consumes:
      - application/json
      description: Memperbarui dosen wali untuk student tertentu. Hanya dapat diakses
        oleh admin dengan permission user:manage
      parameters:
      - description: Student ID
        in: path
        name: id
        required: true
        type: string
      - description: Advisor ID
        in: body
        name: body
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: message
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "422":
          description: Unprocessable Entity
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Update student advisor
      tags:
      - Students
...
schemes:
- http
securityDefinitions:
  Bearer:
    description: 'Type "Bearer" followed by a space and JWT token. Example: "Bearer
      {token}"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"